+++ src/github.com/nlamirault/go-onlinelabs/api/client.go
@@ -1,4 +1,4 @@
-// Copyright (C) 2015  Nicolas Lamirault <nicolas.lamirault@gmail.com>
+// Copyright (C) 2015 Nicolas Lamirault <nicolas.lamirault@gmail.com>
 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,7 +15,6 @@
 package api
 
 import (
-	"bytes"
 	"fmt"
 	"io/ioutil"
 	"net/http"
@@ -37,24 +36,31 @@ type OnlineLabsClient struct {
 	UserID       string
 	Token        string
 	Organization string
-	client       *http.Client
+	Client       *http.Client
+	ComputeURL   string
+	AccountURL   string
 }
 
 // NewClient creates a new OnlineLabs API client using userId, API token and organization
 func NewClient(userid string, token string, organization string) *OnlineLabsClient {
-
 	log.Debugf("Creating client using %s %s %s", userid, token, organization)
 	client := &OnlineLabsClient{
 		UserID:       userid,
 		Token:        token,
 		Organization: organization,
-		client:       &http.Client{},
+		Client:       &http.Client{},
+		ComputeURL:   computeURL,
+		AccountURL:   accountURL,
 	}
 	return client
 }
 
+// GetUserInformations list informations about your user account
 func (c OnlineLabsClient) GetUserInformations(userID string) ([]byte, error) {
-	body, err := c.getAccountAPIResource(fmt.Sprintf("users/%s", userID))
+	body, err := getAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/users/%s", c.AccountURL, userID))
 	if err != nil {
 		return nil, err
 	}
@@ -62,12 +68,93 @@ func (c OnlineLabsClient) GetUserInformations(userID string) ([]byte, error) {
 	return body, nil
 }
 
-func (c OnlineLabsClient) GetUserOrganzations() ([]byte, error) {
-	body, err := c.getAccountAPIResource(fmt.Sprintf("organizations"))
+// GetUserOrganizations list all organizations associate with your account
+func (c OnlineLabsClient) GetUserOrganizations() ([]byte, error) {
+	body, err := getAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/organizations", c.AccountURL))
 	if err != nil {
 		return nil, err
 	}
 	log.Debugf("Get user organizations response: %s", string(body))
+	return body, nil
+}
+
+// GetUserTokens list all tokens associate with your account
+func (c OnlineLabsClient) GetUserTokens() ([]byte, error) {
+	body, err := getAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/tokens", c.AccountURL))
+	if err != nil {
+		return nil, err
+	}
+	log.Debugf("Get tokens response: %s", string(body))
+	return body, nil
+}
+
+//GetUserToken lList an individual Token
+func (c OnlineLabsClient) GetUserToken(tokenID string) ([]byte, error) {
+	body, err := getAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/tokens/%s", c.AccountURL, tokenID))
+	if err != nil {
+		return nil, err
+	}
+	log.Debugf("Get token response: %s", string(body))
+	return body, nil
+}
+
+// CreateToken authenticates a user against their email, password,
+// and then returns a new Token, which can be used until it expires.
+// email is the user email
+// password is the user password
+// expires is if you want a token wich expires or not
+func (c OnlineLabsClient) CreateToken(email string, password string,
+	expires bool) ([]byte, error) {
+	json := fmt.Sprintf(`{"email": "%s", "password": "%s", "expires": %t}`,
+		email, password, expires)
+	body, err := postAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/tokens", c.AccountURL),
+		[]byte(json))
+	if err != nil {
+		return nil, err
+	}
+	log.Debugf("Create token response: %s",
+		string(body))
+	return body, nil
+}
+
+// DeleteToken delete a specific token
+// tokenID ith the token unique identifier
+func (c OnlineLabsClient) DeleteToken(tokenID string) ([]byte, error) {
+	body, err := deleteAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/tokens/%s", c.AccountURL, tokenID))
+	if err != nil {
+		return nil, err
+	}
+	log.Debugf("Delete token response: %s", string(body))
+	return body, nil
+}
+
+// UpdateToken increase Token expiration time of 30 minutes
+// tokenID ith the token unique identifier
+func (c OnlineLabsClient) UpdateToken(tokenID string) ([]byte, error) {
+	body, err := patchAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/tokens/%s", c.AccountURL, tokenID),
+		nil)
+	if err != nil {
+		return nil, err
+	}
+	log.Debugf("Update token response: %s", string(body))
 	return body, nil
 }
 
@@ -79,7 +166,11 @@ func (c OnlineLabsClient) CreateServer(name string, organization string,
 	image string) ([]byte, error) {
 	json := fmt.Sprintf(`{"name": "%s", "organization": "%s", "image": "%s", "tags": ["docker-machine"]}`,
 		name, organization, image)
-	body, err := c.postAPIResource("servers", []byte(json))
+	body, err := postAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/servers", c.ComputeURL),
+		[]byte(json))
 	if err != nil {
 		return nil, err
 	}
@@ -91,7 +182,10 @@ func (c OnlineLabsClient) CreateServer(name string, organization string,
 // DeleteServer delete a specific server
 // serverID ith the server unique identifier
 func (c OnlineLabsClient) DeleteServer(serverID string) ([]byte, error) {
-	body, err := c.deleteAPIResource(fmt.Sprintf("servers/%s", serverID))
+	body, err := deleteAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/servers/%s", c.ComputeURL, serverID))
 	if err != nil {
 		return nil, err
 	}
@@ -103,7 +197,10 @@ func (c OnlineLabsClient) DeleteServer(serverID string) ([]byte, error) {
 // GetServer list an individual server
 // serverID ith the server unique identifier
 func (c OnlineLabsClient) GetServer(serverID string) ([]byte, error) {
-	body, err := c.getAPIResource(fmt.Sprintf("servers/%s", serverID))
+	body, err := getAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/servers/%s", c.ComputeURL, serverID))
 	if err != nil {
 		return nil, err
 	}
@@ -113,7 +210,10 @@ func (c OnlineLabsClient) GetServer(serverID string) ([]byte, error) {
 
 // GetServers list all servers associate with your account
 func (c OnlineLabsClient) GetServers() ([]byte, error) {
-	body, err := c.getAPIResource("servers")
+	body, err := getAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/servers", c.ComputeURL))
 	if err != nil {
 		return nil, err
 	}
@@ -127,14 +227,119 @@ func (c OnlineLabsClient) GetServers() ([]byte, error) {
 // action is the action to execute
 func (c OnlineLabsClient) PerformServerAction(serverID string, action string) ([]byte, error) {
 	json := fmt.Sprintf(`{"action": "%s"}`, action)
-	body, err := c.postAPIResource(
-		fmt.Sprintf("servers/%s/action", serverID),
+	body, err := postAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/servers/%s/action", c.ComputeURL, serverID),
 		[]byte(json))
 	if err != nil {
 		return nil, err
 	}
 	log.Debugf("[OnlineAPI] Server action response: %s",
 		string(body))
+	return body, nil
+}
+
+// GetVolume list an individual volume
+// volumeID ith the volume unique identifier
+func (c OnlineLabsClient) GetVolume(volumeID string) ([]byte, error) {
+	body, err := getAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/volumes/%s", c.ComputeURL, volumeID))
+	if err != nil {
+		return nil, err
+	}
+	log.Debugf("Get volume response: %s", string(body))
+	return body, nil
+}
+
+// DeleteVolume delete a specific volume
+// volumeID ith the volume unique identifier
+func (c OnlineLabsClient) DeleteVolume(volumeID string) ([]byte, error) {
+	body, err := deleteAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/volumes/%s", c.ComputeURL, volumeID))
+	if err != nil {
+		return nil, err
+	}
+	log.Debugf("Delete volume response: %s", string(body))
+	return body, nil
+}
+
+// CreateVolume creates a new volume
+// name is the volume name
+// organization is the organization unique identifier
+// volume_type is the volume type
+// size is the volume size
+func (c OnlineLabsClient) CreateVolume(name string, organization string, volume_type string, size int) ([]byte, error) {
+	json := fmt.Sprintf(`{"name": "%s", "organization": "%s", "volume_type": "%s", "size": %d}`,
+		name, organization, volume_type, size)
+	body, err := postAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/volumes", c.ComputeURL),
+		[]byte(json))
+	if err != nil {
+		return nil, err
+	}
+	log.Debugf("Create volume response: %s", string(body))
+	return body, nil
+}
+
+// GetVolumes list all volumes associate with your account
+func (c OnlineLabsClient) GetVolumes() ([]byte, error) {
+	body, err := getAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/volumes", c.ComputeURL))
+	if err != nil {
+		return nil, err
+	}
+	log.Debugf("Retrieve volumes response: %s",
+		string(body))
+	return body, nil
+}
+
+// GetImages list all images associate with your account
+func (c OnlineLabsClient) GetImages() ([]byte, error) {
+	body, err := getAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/images", c.ComputeURL))
+	if err != nil {
+		return nil, err
+	}
+	log.Debugf("Retrieve images response: %s", string(body))
+	return body, nil
+}
+
+// GetImage list an individual image
+// volumeID ith the image unique identifier
+func (c OnlineLabsClient) GetImage(volumeID string) ([]byte, error) {
+	body, err := getAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/images/%s", c.ComputeURL, volumeID))
+	if err != nil {
+		return nil, err
+	}
+	log.Debugf("Get image response: %s", string(body))
+	return body, nil
+}
+
+// DeleteImage delete a specific volume
+// volumeID ith the volume unique identifier
+func (c OnlineLabsClient) DeleteImage(imageID string) ([]byte, error) {
+	body, err := deleteAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/images/%s", c.ComputeURL, imageID))
+	if err != nil {
+		return nil, err
+	}
+	log.Debugf("Delete image response: %s", string(body))
 	return body, nil
 }
 
@@ -148,8 +353,10 @@ func (c OnlineLabsClient) UploadPublicKey(userid string, keyPath string) ([]byte
 	}
 	json := fmt.Sprintf(`{"ssh_public_keys": [{"key": "%s"}]}`,
 		strings.TrimSpace(string(publicKey)))
-	body, err := c.patchAPIResource(
-		fmt.Sprintf("%s/users/%s", accountURL, userid),
+	body, err := patchAPIResource(
+		c.Client,
+		c.Token,
+		fmt.Sprintf("%s/users/%s", c.AccountURL, userid),
 		[]byte(json))
 	if err != nil {
 		return nil, err
@@ -157,104 +364,4 @@ func (c OnlineLabsClient) UploadPublicKey(userid string, keyPath string) ([]byte
 	log.Debugf("[OnlineAPI] Server action response: %s",
 		string(body))
 	return body, nil
-}
-
-func (c OnlineLabsClient) getAccountAPIResource(request string) ([]byte, error) {
-	url := fmt.Sprintf("%s/%s", accountURL, request)
-	log.Debugf("GET: %q", url)
-	req, err := http.NewRequest("GET", url, nil)
-	req.Header.Set("X-Auth-Token", c.Token)
-	req.Header.Set("Content-Type", "application/json")
-	resp, err := c.client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	b, err := ioutil.ReadAll(resp.Body)
-	if resp.StatusCode > 299 {
-		return nil, fmt.Errorf("Status code: %d", resp.StatusCode)
-	}
-	return b, nil
-}
-
-func (c OnlineLabsClient) getAPIResource(request string) ([]byte, error) {
-	url := fmt.Sprintf("%s/%s", computeURL, request)
-	log.Debugf("GET: %q", url)
-	req, err := http.NewRequest("GET", url, nil)
-	req.Header.Set("X-Auth-Token", c.Token)
-	req.Header.Set("Content-Type", "application/json")
-	resp, err := c.client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	b, err := ioutil.ReadAll(resp.Body)
-	if resp.StatusCode > 299 {
-		return nil, fmt.Errorf("Status code: %d", resp.StatusCode)
-	}
-	return b, nil
-}
-
-func (c OnlineLabsClient) postAPIResource(request string, json []byte) ([]byte, error) {
-	url := fmt.Sprintf("%s/%s", computeURL, request)
-	log.Debugf("POST: %q %s", url, string(json))
-	req, err := http.NewRequest("POST", url, bytes.NewBuffer(json))
-	req.Header.Set("X-Auth-Token", c.Token)
-	req.Header.Set("Content-Type", "application/json")
-	resp, err := c.client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	b, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-	if resp.StatusCode > 299 {
-		return nil, fmt.Errorf("%d %s",
-			resp.StatusCode, string(b))
-	}
-	//return ioutil.ReadAll(resp.Body)
-	return b, nil
-}
-
-func (c OnlineLabsClient) deleteAPIResource(request string) ([]byte, error) {
-	url := fmt.Sprintf("%s/%s", computeURL, request)
-	log.Debugf("DELETE: %q", url)
-	req, err := http.NewRequest("DELETE", url, nil)
-	req.Header.Set("X-Auth-Token", c.Token)
-	req.Header.Set("Content-Type", "application/json")
-	resp, err := c.client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	b, err := ioutil.ReadAll(resp.Body)
-	if resp.StatusCode > 299 {
-		return nil, fmt.Errorf("Status code: %d", resp.StatusCode)
-	}
-	return b, nil
-}
-
-func (c OnlineLabsClient) patchAPIResource(url string, json []byte) ([]byte, error) {
-	//url := fmt.Sprintf("%s/%s", computeURL, request)
-	log.Debugf("PATCH: %q %s", url, string(json))
-	req, err := http.NewRequest("PATCH", url, bytes.NewBuffer(json))
-	req.Header.Set("X-Auth-Token", c.Token)
-	req.Header.Set("Content-Type", "application/json")
-	resp, err := c.client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	b, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-	if resp.StatusCode > 299 {
-		return nil, fmt.Errorf("%d %s",
-			resp.StatusCode, string(b))
-	}
-	//return ioutil.ReadAll(resp.Body)
-	return b, nil
 }