+++ src/github.com/nlamirault/go-onlinelabs/api/server.go
@@ -17,6 +17,8 @@ package api
 import (
 	"encoding/json"
 	//"fmt"
+
+	log "github.com/Sirupsen/logrus"
 )
 
 // PublicIP represents IP address of a server
@@ -52,7 +54,7 @@ type ServersResponse struct {
 // GetServerFromJSON load bytes and return a ServerResponse
 func GetServerFromJSON(b []byte) (*ServerResponse, error) {
 	var response ServerResponse
-	// fmt.Printf("Response JSON: %s\n", (string(b)))
+	//fmt.Printf("Response JSON: %s\n", (string(b)))
 	err := json.Unmarshal(b, &response)
 	if err != nil {
 		return nil, err
@@ -68,4 +70,12 @@ func GetServersFromJSON(b []byte) (*ServersResponse, error) {
 		return nil, err
 	}
 	return &response, nil
+}
+
+func (s Server) Display() {
+	log.Infof("Id   : %s", s.ID)
+	log.Infof("Name : %s", s.Name)
+	log.Infof("Date : %s", s.ModificationDate)
+	log.Infof("IP   : %s", s.PublicIP.Address)
+	log.Infof("Tags : %s", s.Tags)
 }