+++ src/github.com/nlamirault/go-onlinelabs/api/user.go
@@ -17,6 +17,8 @@ package api
 import (
 	"encoding/json"
 	//"fmt"
+
+	log "github.com/Sirupsen/logrus"
 )
 
 // Organization represents an Online Labs entity
@@ -56,6 +58,25 @@ type UserResponse struct {
 	User User `json:"user,omitempty"`
 }
 
+// Token represents an identifier associated with your account.
+// It is used to authenticate commands in the APIs
+type Token struct {
+	ID       string `json:"id,omitempty"`
+	UserID   string `json:"user_id,omitempty"`
+	Creation string `json:"creation_date,omitempty"`
+	Expires  string `json:"expires,omitempty"`
+}
+
+// TokenResponse represents JSON response of server for a token
+type TokenResponse struct {
+	Token Token `json:"token,omitempty"`
+}
+
+// TokensResponse represents JSON response of server for tokens
+type TokensResponse struct {
+	Tokens []Token
+}
+
 // GetUserFromJSON load bytes and return a UserResponse
 func GetUserFromJSON(b []byte) (*UserResponse, error) {
 	var response UserResponse
@@ -67,7 +88,7 @@ func GetUserFromJSON(b []byte) (*UserResponse, error) {
 	return &response, nil
 }
 
-// GetUserFromJSON load bytes and return a UserResponse
+// GetOrganizationsFromJSON load bytes and return a OrganizationsResponse
 func GetOrganizationsFromJSON(b []byte) (*OrganizationsResponse, error) {
 	var response OrganizationsResponse
 	// fmt.Printf("Response JSON: %s\n", (string(b)))
@@ -76,4 +97,52 @@ func GetOrganizationsFromJSON(b []byte) (*OrganizationsResponse, error) {
 		return nil, err
 	}
 	return &response, nil
+}
+
+// GetTokenFromJSON load bytes and return a TokenResponse
+func GetTokenFromJSON(b []byte) (*TokenResponse, error) {
+	var response TokenResponse
+	// fmt.Printf("Response JSON: %s\n", (string(b)))
+	err := json.Unmarshal(b, &response)
+	if err != nil {
+		return nil, err
+	}
+	return &response, nil
+}
+
+// GetTokensFromJSON load bytes and return a TokensResponse
+func GetTokensFromJSON(b []byte) (*TokensResponse, error) {
+	var response TokensResponse
+	// fmt.Printf("Response JSON: %s\n", (string(b)))
+	err := json.Unmarshal(b, &response)
+	if err != nil {
+		return nil, err
+	}
+	return &response, nil
+}
+
+func (u User) Display() {
+	log.Infof("Id            : %s", u.ID)
+	log.Infof("Fullname      : %s", u.Fullname)
+	log.Infof("Firstname     : %s", u.Firstname)
+	log.Infof("Lastname      : %s", u.Lastname)
+	log.Infof("Email         : %s", u.Email)
+	log.Infof("Phone         : %s", u.PhoneNumber)
+	log.Infof("Roles         : %s", u.Roles)
+	log.Infof("Organizations : %s", u.Organizations)
+}
+
+func (o Organization) Display() {
+	log.Infof("Id              : %s", o.ID)
+	log.Infof("Name            : %s", o.Name)
+	log.Infof("Currency        : %s", o.Currency)
+	log.Infof("Locale          : %s", o.Locale)
+	log.Infof("Customer class  : %s", o.CustomerClass)
+}
+
+func (t Token) Display() {
+	log.Infof("Id        : %s", t.ID)
+	log.Infof("UserId    : %s", t.UserID)
+	log.Infof("Creation  : %s", t.Creation)
+	log.Infof("Expires   : %s", t.Expires)
 }