+++ src/github.com/nlamirault/go-onlinelabs/main_test.go
@@ -31,19 +31,30 @@ func checkCommand(t *testing.T, name string, command cli.Command) {
 
 func TestCLICommands(t *testing.T) {
 	app := makeApp()
-	if len(app.Commands) != 7 {
-		t.Errorf("Invalid CLI commands")
+	if len(app.Commands) != 18 {
+		t.Errorf("Invalid CLI number of commands")
 	}
 	// for _, command := range app.Commands {
 	// 	fmt.Printf("command : %v", command)
 	// }
-	checkCommand(t, "listImages", app.Commands[0])
-	checkCommand(t, "listServers", app.Commands[1])
-	checkCommand(t, "getServer", app.Commands[2])
-	checkCommand(t, "deleteServer", app.Commands[3])
-	checkCommand(t, "actionServer", app.Commands[4])
-	checkCommand(t, "getUser", app.Commands[5])
-	checkCommand(t, "getOrganizations", app.Commands[6])
+	checkCommand(t, "listServers", app.Commands[0])
+	checkCommand(t, "getServer", app.Commands[1])
+	checkCommand(t, "deleteServer", app.Commands[2])
+	checkCommand(t, "actionServer", app.Commands[3])
+	checkCommand(t, "getUser", app.Commands[4])
+	checkCommand(t, "getOrganizations", app.Commands[5])
+	checkCommand(t, "getTokens", app.Commands[6])
+	checkCommand(t, "getToken", app.Commands[7])
+	checkCommand(t, "deleteToken", app.Commands[8])
+	checkCommand(t, "createToken", app.Commands[9])
+	checkCommand(t, "updateToken", app.Commands[10])
+	checkCommand(t, "listVolumes", app.Commands[11])
+	checkCommand(t, "getVolume", app.Commands[12])
+	checkCommand(t, "deleteVolume", app.Commands[13])
+	checkCommand(t, "createVolume", app.Commands[14])
+	checkCommand(t, "listImages", app.Commands[15])
+	checkCommand(t, "getImage", app.Commands[16])
+	checkCommand(t, "deleteImage", app.Commands[17])
 }
 
 func checkGlobalArgument(flags []cli.Flag, name string) int {